# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MICM Manufacturing Maturity Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessment cycles
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of assessment cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentCycle"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create a new assessment cycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentBody"
      responses:
        "201":
          description: Created assessment cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentCycle"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get an assessment cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentCycle"
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update an assessment cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssessmentBody"
      responses:
        "200":
          description: Updated assessment cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentCycle"
  /assessments/{id}/assign:
    post:
      operationId: assignAssessment
      tags: [assessments]
      summary: Assign users to an assessment cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAssessmentBody"
      responses:
        "200":
          description: Assigned users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentCycle"
  /assessments/{id}/results:
    get:
      operationId: getAssessmentResults
      tags: [assessments]
      summary: Get aggregated results for an assessment cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment results with per-user and aggregate scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResults"

  # ─── Scores ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AssessmentCycle:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, completed]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        assignedUserIds:
          type: array
          items:
            type: integer
        completedUserIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - companyId
        - name
        - status
        - assignedUserIds
        - completedUserIds
        - createdAt
        - updatedAt
    CreateAssessmentBody:
      type: object
      properties:
        companyId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
      required:
        - companyId
        - name
    UpdateAssessmentBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    AssignAssessmentBody:
      type: object
      properties:
        userIds:
          type: array
          items:
            type: integer
      required:
        - userIds
    AssessmentResults:
      type: object
      properties:
        assessmentId:
          type: integer
        assessmentName:
          type: string
        userScores:
          type: array
          items:
            $ref: "#/components/schemas/UserScoreSet"
        aggregateScores:
          type: array
          items:
            $ref: "#/components/schemas/DomainScore"
      required:
        - assessmentId
        - assessmentName
        - userScores
        - aggregateScores
    UserScoreSet:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        domainScores:
          type: array
          items:
            $ref: "#/components/schemas/DomainScore"
        overallScore:
          type: ["number", "null"]
        completedAt:
          type: ["string", "null"]
      required:
        - userId
        - userName
        - domainScores
    DomainScore:
      type: object
      properties:
        domainId:
          type: integer
        domainName:
          type: string
        score:
          type: ["number", "null"]
        band:
          type: ["string", "null"]
      required:
        - domainId
        - domainName

    # ─── Score ───────────────────────────────────────────────────────────────────
...